Exclusive neri branch

+++ app/src/main/java/com/example/nixapp/DB/Pagos.java
@@ -3,13 +3,13 @@ package com.example.nixapp.DB;
 public class Pagos {
     String paymentMethodId;
     String descripcion;
-    int monto;
+    double monto;
 
     public Pagos(String paymentMethodId) {
         this.paymentMethodId = paymentMethodId;
     }
 
-    public Pagos(String paymentMethodId, String descripcion, int monto) {
+    public Pagos(String paymentMethodId, String descripcion, double monto) {
         this.paymentMethodId = paymentMethodId;
         this.descripcion = descripcion;
         this.monto = monto;
